requirements.txt
pygame-ce>=2.5,<3.0
pymunk>=6.8,<7.0
